Split backend agent API, settings API, and agent run service
Move agent and settings routers into directory modules; extract MCP
policy helpers and shared skill path/registry utilities; split
agentRunService into agent-runs types, interrupts, persistence, and
lifecycle with a compatibility barrel.
