+++ backend/src/services/skills/constants.ts
@@ -1,0 +1,7 @@
+import path from 'path';
+import { existsSync } from 'fs';
+
+/** Repository root (parent of `backend/`) */
+export const repoRoot = path.resolve(__dirname, '../../../..');
+
+export const skillsRoot = process.env.SKILLS_ROOT || path.join(repoRoot, 'skills');
